requirements.txt
Django==1.10.2
mysqlclient==1.3.9
django-debug-toolbar==1.6
django-mailer==1.2.2
django-allauth==0.27.0
django-last-seen==0.3
django-timezone-field==2.0
pytz